Move the setup commands to a dedicated namespace
