+++ src/command_line/setup/jdk_command.cs
@@ -1,4 +1,4 @@
-namespace Belin.Cli.CommandLine;
+namespace Belin.Cli.CommandLine.Setup;
 
 /// <summary>
 /// Downloads and installs the latest OpenJDK release.
