+++ src/command_line/setup/node_command.coffee
